AuditPro_SHARE/requirements.txt
# ── AuditPro — dépendances complètes ─────────────────────────────
# Interface graphique
PyQt6>=6.5.0

# Traitement de données
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1          # Support fichiers .xls (ancien format Excel)
lxml>=4.9.0          # Requis par python-docx pour XML avancé

# Génération Word
python-docx>=1.0.0

# PDF
pdfplumber>=0.10.0
pdf2image>=1.16.0
pypdf>=4.0.0         # Split PDF (circularisation étape 3)
pytesseract>=0.3.10  # OCR (nécessite Tesseract-OCR installé séparément)

# Images / Icônes
Pillow>=10.0.0       # Génération icône .ico + traitements images

# Agent IA
ollama>=0.3.0           # Client Ollama pour l'agent Llama local
fastapi>=0.111.0        # API REST pour exposer l'agent AuditPro
uvicorn>=0.30.0         # Serveur ASGI pour lancer FastAPI

# Packaging
pyinstaller>=6.0.0
